+++ packages/koni_jxl/lib/koni_jxl.dart
@@ -3,5 +3,5 @@ library;
 
 export 'src/decoder.dart' show JxlDecoder;
 export 'src/exceptions.dart';
-export 'src/jxl_image.dart' show JxlImage;
+export 'src/jxl_image.dart' show JxlAnimation, JxlImage;
 export 'src/jxl_info.dart' show JxlInfo;