+++ packages/koni_jxl_flutter/lib/koni_jxl_flutter.dart
@@ -3,5 +3,6 @@ library;
 
 export 'package:koni_jxl/koni_jxl.dart';
 
+export 'src/jxl_animation_view.dart';
 export 'src/jxl_codec.dart';
 export 'src/jxl_image_provider.dart';